requirements.txt
absl-py==0.6.1
alabaster==0.7.10
alyn==0.1.1
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
awscli==1.16.135
awsebcli==3.14.13
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
blessed==1.15.0
bokeh==0.12.16
boto==2.48.0
boto3==1.9.36
botocore==1.12.125
Bottleneck==1.2.1
bz2file==0.98
cached-property==1.5.1
cachetools==3.1.0
cement==2.8.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
conda==4.5.12
conda-build==3.10.5
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.2.2
cycler==0.10.0
cymem==1.31.2
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
deepspeech==0.4.0
dill==0.2.8.2
distributed==1.21.8
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ebcli==4.0.18
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
Fabric3==1.14.post1
fastcache==1.0.2
ffmpeg==1.4
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
funcy==1.11
future==0.16.0
gast==0.2.1.post0
gensim==3.6.0
gevent==1.3.0
glob2==0.6
gmpy2==2.0.8
google-api-core==1.8.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-images-download==2.5.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
greenlet==0.4.13
grpcio==1.14.1
gunicorn==19.9.0
gym==0.10.5
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.3.0
imagesize==1.0.0
imbalanced-learn==0.4.3
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.1
JSAnimation==0.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
kazoo==2.5.0
Keras==2.2.4
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
keyring==9.0
keyrings.alt==3.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.5.6
multipledispatch==0.5.0
murmurhash==0.26.4
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.14.3
numpydoc==0.8.0
oauthlib==3.0.0
odo==0.5.1
olefile==0.45.1
opencv-python==3.4.1.15
openpyxl==2.5.3
packaging==17.1
pandas==0.23.0
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
pathspec==0.5.9
patsy==0.5.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
plac==1.0.0
pluggy==0.6.0
ply==3.11
pocketsphinx==0.1.15
preshed==0.46.4
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.5
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.4
pyasn1-modules==0.2.4
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydot==1.4.1
pydub==0.23.0
pyflakes==1.6.0
pyglet==1.2.4
Pygments==2.2.0
pygpu==0.7.6
pykafka==2.8.0
pyLDAvis==2.1.2
pylint==1.8.4
PyNaCl==1.3.0
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pystorm==3.1.4
pytesseract==0.2.2
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
python-pptx==0.6.17
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
requests==2.20.1
requests-oauthlib==1.0.0
rope==0.10.7
rsa==3.4.2
ruamel-yaml==0.15.35
ruamel.yaml==0.15.72
s3transfer==0.2.0
scikit-image==0.13.1
scikit-learn==0.20.2
scipy==1.1.0
seaborn==0.8.1
selenium==3.141.0
semantic-version==2.5.0
semver==2.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.7.1
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
spacy==0.101.0
SpeechRecognition==3.6.3
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
sputnik==0.9.3
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.9.0
streamparse==3.14.0
sympy==1.1.1
tables==3.4.3
tabulate==0.8.2
tblib==1.3.2
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
tesseract==0.1.3
testpath==0.3.1
textblob==0.15.1
texttable==0.9.1
Theano==1.0.3
thinc==5.0.8
thriftpy==0.3.9
toolz==0.9.0
tornado==5.0.2
tqdm==4.31.1
traitlets==4.3.2
tweepy==3.7.0
typing==3.6.4
ujson==1.35
unicodecsv==0.14.1
urllib3==1.24.1
vaderSentiment==3.2.1
virtualenv==16.4.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
xgboost==0.81
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
zict==0.1.3
